+++ anybooru/__init__.py
@@ -4,7 +4,7 @@
 Anybooru
 
 Anybooru is an API client for Danbooru, Moebooru, Serika, e621ng, Zerochan,
-Gelbooru, e-shuushuu, Gelbooru 0.2 based sites and the Sakuria API.
+Gelbooru, e-shuushuu, Gelbooru 0.2 based sites, Sakuria and Anime-Pictures APIs.
 
 Anybooru requires "requests" package to work.
 
@@ -19,6 +19,7 @@ Anybooru modules:
     shuushuu -- Contains Shuushuu main class.
     gelbooru02 -- Contains Gelbooru02 main class for Gelbooru 0.2 sites.
     sakuria -- Contains Sakuria main class for the Sakuria API.
+    anime_pictures -- Contains AnimePictures main class.
     api_moebooru -- Contains all Moebooru API functions.
     api_danbooru -- Contains all Danbooru API functions.
     api_serika -- Contains official v1 and internal Serika API functions.
@@ -28,6 +29,7 @@ Anybooru modules:
     api_shuushuu -- Contains native e-shuushuu API functions.
     api_gelbooru02 -- Contains native Gelbooru 0.2 API functions.
     api_sakuria -- Contains native Sakuria API functions.
+    api_anime_pictures -- Contains native Anime-Pictures API functions.
     exceptions -- Manages and builds Anybooru errors messages.
     resources -- Packaged default parameters (DEFAULT_CONFIG_FILE) and encoding.
 """
@@ -47,5 +49,6 @@ from .gelbooru import Gelbooru
 from .shuushuu import Shuushuu
 from .gelbooru02 import Gelbooru02
 from .sakuria import Sakuria
+from .anime_pictures import AnimePictures
 from .exceptions import (AnybooruError, AnybooruAPIError, AnybooruHTTPError)
 from .resources import DEFAULT_CONFIG_FILE